Generate integer types for counts from a YAML override table
The TypeScript schema types token counts, byte limits, document
versions and timestamps as a bare number, so they came out as float64,
and callers wrote new(float64(limit)) for a byte limit. The protocol's
Rust schema declares every one of them an integer.

internal/cmd/schema/overrides.yaml now lists such members with their Go
integer type, as the Rust schema's format gives it. The generator
embeds the file, rejects unknown keys and types, applies each entry to
every version that has the member, and fails when an entry names a
member no version has or one that is not a number. SendUsage takes
uint64 to match.

+++ acp1/session_stream.go
@@ -207,7 +207,7 @@ func (s *SessionStream) SendCommands(ctx context.Context, commands []AvailableCo
 
 // SendUsage reports context window usage for the turn so far: used tokens out
 // of size, with an optional running cost.
-func (s *SessionStream) SendUsage(ctx context.Context, used, size float64, cost *Cost) error {
+func (s *SessionStream) SendUsage(ctx context.Context, used, size uint64, cost *Cost) error {
 	return s.Send(ctx, schema.SessionUpdateUsageUpdate{
 		Used: used,
 		Size: size,
