+++ schema/v2/types.gen.go
@@ -996,18 +996,18 @@ type ContentChunk struct {
 // Experimental: not part of the spec yet; it may change or be removed.
 type Usage struct {
 	// Sum of all token types across session.
-	TotalTokens float64 `json:"totalTokens"`
+	TotalTokens uint64 `json:"totalTokens"`
 	// Total input tokens.
-	InputTokens float64 `json:"inputTokens"`
+	InputTokens uint64 `json:"inputTokens"`
 	// Total output tokens.
-	OutputTokens float64 `json:"outputTokens"`
+	OutputTokens uint64 `json:"outputTokens"`
 	// Total thought/reasoning tokens
-	ThoughtTokens *float64 `json:"thoughtTokens,omitzero"`
+	ThoughtTokens *uint64 `json:"thoughtTokens,omitzero"`
 	// Total cache read tokens.
-	CachedReadTokens *float64 `json:"cachedReadTokens,omitzero"`
+	CachedReadTokens *uint64 `json:"cachedReadTokens,omitzero"`
 	// Total cache write tokens.
-	CachedWriteTokens *float64 `json:"cachedWriteTokens,omitzero"`
-	Meta              Meta     `json:"_meta,omitzero"`
+	CachedWriteTokens *uint64 `json:"cachedWriteTokens,omitzero"`
+	Meta              Meta    `json:"_meta,omitzero"`
 }
 
 // TerminalOutput is an authoritative replacement snapshot of terminal output bytes.
@@ -1478,7 +1478,7 @@ type SuggestNesRequest struct {
 	// The URI of the document to suggest for.
 	URI string `json:"uri"`
 	// The version number of the document.
-	Version float64 `json:"version"`
+	Version int64 `json:"version"`
 	// The current cursor position.
 	Position Position `json:"position"`
 	// The current text selection range, if any.
@@ -1556,8 +1556,8 @@ type NesUserAction struct {
 	// The position where the action occurred.
 	Position Position `json:"position"`
 	// Timestamp in milliseconds since epoch.
-	TimestampMs float64 `json:"timestampMs"`
-	Meta        Meta    `json:"_meta,omitzero"`
+	TimestampMs uint64 `json:"timestampMs"`
+	Meta        Meta   `json:"_meta,omitzero"`
 }
 
 // NesOpenFile is an open file in the editor.
@@ -1569,8 +1569,8 @@ type NesOpenFile struct {
 	// The visible range in the editor, if any.
 	VisibleRange *Range `json:"visibleRange,omitzero"`
 	// Timestamp in milliseconds since epoch of when the file was last focused.
-	LastFocusedMs *float64 `json:"lastFocusedMs,omitzero"`
-	Meta          Meta     `json:"_meta,omitzero"`
+	LastFocusedMs *uint64 `json:"lastFocusedMs,omitzero"`
+	Meta          Meta    `json:"_meta,omitzero"`
 }
 
 // NesDiagnostic is a diagnostic (error, warning, etc.).
@@ -1645,7 +1645,7 @@ type DidOpenDocumentNotification struct {
 	// The language identifier of the document (e.g., "rust", "python").
 	LanguageID string `json:"languageId"`
 	// The version number of the document.
-	Version float64 `json:"version"`
+	Version int64 `json:"version"`
 	// The full text content of the document.
 	Text string `json:"text"`
 	Meta Meta   `json:"_meta,omitzero"`
@@ -1658,7 +1658,7 @@ type DidChangeDocumentNotification struct {
 	// The URI of the changed document.
 	URI string `json:"uri"`
 	// The new version number of the document.
-	Version float64 `json:"version"`
+	Version int64 `json:"version"`
 	// The content changes.
 	ContentChanges []TextDocumentContentChangeEvent `json:"contentChanges"`
 	Meta           Meta                             `json:"_meta,omitzero"`
@@ -1701,7 +1701,7 @@ type DidFocusDocumentNotification struct {
 	// The URI of the focused document.
 	URI string `json:"uri"`
 	// The version number of the document.
-	Version float64 `json:"version"`
+	Version int64 `json:"version"`
 	// The current cursor position.
 	Position Position `json:"position"`
 	// The portion of the file currently visible in the editor viewport.
